hybris: tests: Use shared hwcomposer code for all tests
