+++ hybris/tests/test_glesv2.cpp
@@ -23,6 +23,8 @@
 #include <stddef.h>
 #include <stdlib.h>
 
+#include "test_common.h"
+
 const char vertex_src [] =
 "                                        \
    attribute vec4        position;       \
@@ -49,16 +51,6 @@ const char fragment_src [] =
              + atan(pos.y,pos.x) - phase );            \
    }                                                   \
 ";
-
-GLuint load_shader(const char *shader_source, GLenum type)
-{
-	GLuint  shader = glCreateShader(type);
-
-	glShaderSource(shader, 1, &shader_source, NULL);
-	glCompileShader(shader);
-
-	return shader;
-}
 
 
 GLfloat norm_x    =  0.0;
@@ -94,14 +86,21 @@ int main(int argc, char **argv)
 	};
 	EGLSurface surface;
 	EGLint ctxattr[] = {
-	EGL_CONTEXT_CLIENT_VERSION, 2,
+		EGL_CONTEXT_CLIENT_VERSION, 2,
 		EGL_NONE
 	};
 	EGLContext context;
 
 	EGLBoolean rv;
 
-	display = eglGetDisplay(NULL);
+	HWComposer *win = create_hwcomposer_window();
+
+	if (!win) {
+		printf("Failed to create native window\n");
+		return 1;
+	}
+
+	display = eglGetDisplay(EGL_DEFAULT_DISPLAY);
 	assert(eglGetError() == EGL_SUCCESS);
 	assert(display != EGL_NO_DISPLAY);
 
@@ -113,7 +112,8 @@ int main(int argc, char **argv)
 	assert(eglGetError() == EGL_SUCCESS);
 	assert(rv == EGL_TRUE);
 
-	surface = eglCreateWindowSurface((EGLDisplay) display, ecfg, (EGLNativeWindowType)NULL, NULL);
+	surface = eglCreateWindowSurface((EGLDisplay) display, ecfg,
+		(EGLNativeWindowType) static_cast<ANativeWindow *> (win), NULL);
 	assert(eglGetError() == EGL_SUCCESS);
 	assert(surface != EGL_NO_SURFACE);
 
@@ -127,15 +127,7 @@ int main(int argc, char **argv)
 	assert(version);
 	printf("%s\n",version);
 
-
-	GLuint vertexShader   = load_shader ( vertex_src , GL_VERTEX_SHADER  );     // load vertex shader
-	GLuint fragmentShader = load_shader ( fragment_src , GL_FRAGMENT_SHADER );  // load fragment shader
-
-	GLuint shaderProgram  = glCreateProgram ();                 // create program object
-	glAttachShader ( shaderProgram, vertexShader );             // and attach both...
-	glAttachShader ( shaderProgram, fragmentShader );           // ... shaders to it
-
-	glLinkProgram ( shaderProgram );    // link the program
+	GLuint shaderProgram = create_program(vertex_src, fragment_src);
 	glUseProgram  ( shaderProgram );    // and select it for usage
 
 	//// now get the locations (kind of handle) of the shaders variables
